# improved TextAnalyzer speed