What's done:
 - speed improved by 6,57s
 - implemented unit test for TextAnalyzer

Benchmark before improvement:
WordCounterBench::benchCounter() - 1,635,216b - 6,881,077.000μs
WordCounterBench::benchCounterWithWordsExport() - 2,112,968b - 6,858,021.000μs

Benchmark after improvement:
WordCounterBench::benchCounter() - 1,740,544b - 312,466.000μs
WordCounterBench::benchCounterWithWordsExport() - 2,115,576b - 371,145.000μs